Move ANSI interpreter to TermInputText
